requirements.txt
pyfaidx>=0.7.2
biopython>=1.81
mygene>=3.2.2
requests>=2.31.0
